+++ src/main/resources/templates/index.hbs
@@ -46,7 +46,7 @@
 </table>
 
 <form method="post" action="/new">
-    <input type="text" name="name">
+    <input type="text" name="name" required>
     <input type="submit" value="방 만들기">
 </form>
